Add files via upload
Added class Hand with __init__ method that initializes the card hand as an empty list, the value of the card hand as zero, and the number of aces at zero. The add_card() method adds a card to the hand by taking in the Deck().deal_card() that adds one Card object (Card(suit, rank)) to the hand. The value of this Card object is retrieved from the global variable values dictionary and updates the value of the hand from each added card.