Add files via upload
I added game_active = True as the first line inside the outer while loop to make sure when the player decides to play again, the game_active variable refers to True so the game will ask the player if they want to hit or stand. Further, I added the call to the ace_adjust() method twice to make sure the game adjusts the dealer's or player's hand based on the presence of aces.